restore soft links
